refactor : 상수 폴더 Depth변경
