+++ constants/competitionDetail.ts
